+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- New utility script to remove unfinished jobs from a results directory: `parman-clean-results`.
+
 ## [0.3.1] - 2023-06-15
 
 ### Fixed